move tag test to seperate file
